DASH-434 Move ClickEventTypes into main type.d.ts file
