--- client/src/pages/SettingsPage/components/ChangePin/types.d.ts
@@ -1,5 +1,0 @@
-import React from 'react';
-
-export type ClickEventTypes = React.MouseEvent<HTMLElement, MouseEvent> & {
-  target: { blur: () => void };
-};
